Fixed a bug in Buffer.hpp's special member functions that caused move to crash.
Fixed bug in Allocator that made it try to map Device Local memory.

Fixed bug in QueueFamilies that made it not return the correct QueueIndex from its get functions.

Fixed bug in VK_CHECK that made it not trigger int3

Fixed bug in FrameHandler that made it so the CommandPool for the transfer queue is now created properly using the transfer queue.

Example transfer in Graphics::draw().

Renderer::render_frame now optionally waits for TransferEpoch semaphore to finnish.

TransferManager now acctually uses its CommandBuffer.

VulkanContext now owns Allocator

+++ src/Engine/src/graphics/vk/resource/GeometryBuffer.cpp
@@ -4,7 +4,7 @@
 #include "GeometryBuffer.hpp"
 namespace odin::graphics::vk::resource
 {
-GeometryBuffer::GeometryBuffer(std::uint32_t size)
+GeometryBuffer::GeometryBuffer(std::uint64_t size)
     : m_BufferSize{ size }
     , m_TransferOffset{ 0u } {};
 GeometryBuffer::GeometryBuffer(GeometryBuffer&& other) noexcept
